requirements.txt
Jinja2==2.7.2
MarkupSafe==0.23
Pillow==2.4.0
PyPDF2==1.21
Pygments==1.6
Sphinx==1.2.2
cov-core==1.12
coverage==3.7.1
docutils==0.11
html5lib==0.999
markdown2==2.2.1
pkginfo==1.1
py==1.4.20
pysqlite==2.6.3
pytest==2.5.2
pytest-cov==1.6
python-dateutil==2.2
python-gnupg==0.3.6
reportlab==3.1.8
requests==2.3.0
six==1.6.1
tox==1.7.1
twine==1.3.1
virtualenv==1.11.5
wheel==0.23.0
wsgiref==0.1.2
xhtml2pdf==0.0.6
